+++ build.gradle
@@ -32,6 +32,5 @@ repositories {
 }
 
 application {
-    // Change this to your main class.
     mainClassName = "Duke"
 }
