+++ src/main/java/command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package command;
+
+import exception.DukeException;
+import task.Task;
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
+
 /**
  * The DeleteCommand class is used when the user intends to delete a particular task from their task list.
  *
