+++ src/main/java/task/Deadline.java
@@ -1,15 +1,16 @@
+package task;
 import java.io.Serializable;
 /**
- * This extension of the task class will allow the user to add a task of event type.
+ * This extension of the task class will allow the user to add a task of deadline type.
  *
  * @author Sai Ganesh Suresh
  * @version v2.0
  */
-public class Event extends Task implements Serializable{
+public class Deadline extends Task implements Serializable{
 
     private String date;
 
-    public Event(String description, String date) {
+    public Deadline(String description, String date){
         super(description);
         this.date = date;
     }
@@ -22,6 +23,7 @@ public class Event extends Task implements Serializable{
      */
     @Override
     public String toString() {
-        return "[E]" +  "[" + super.getStatusIcon() + "]" + this.description + "(at: " + date + ")";
+        return "[D]" + "[" + super.getStatusIcon() + "]" + this.description + "(by: " + date + ")";
     }
 }
+
