+++ src/main/java/task/TaskList.java
@@ -1,3 +1,4 @@
+package task;
 import java.util.ArrayList;
 /**
  * The TaskList class handles all operations performed on the TaskList as well as stores the TaskList.
