+++ src/main/java/task/Todo.java
@@ -1,3 +1,4 @@
+package task;
 import java.io.Serializable;
 /**
  * This extension of the task class will allow the user to add a task of to-do type.
@@ -19,6 +20,6 @@ public class Todo extends Task implements Serializable {
      */
     @Override
     public String toString() {
-        return "[T]" + "[" + super.getStatusIcon() + "]" + this.description;
+        return "[T]" + "[" + super.getStatusIcon() + "] " + this.description;
     }
 }
