--- src/main/text-ui-testing/EXPECTED.txt
@@ -1,16 +1,0 @@
-_______________________________
-Hello! I'm Duke
-What can I do for you?
-_______________________________
-
-_______________________________
-Here are the task(s) in your list:
-1.[E][?]this(at: Tue Feb 12 23:00:00 GMT+08:00 2019)
-2.[D][?]return books(by: Sat Feb 20 13:00:00 GMT+08:00 1999)
-3.[T][?]this
-4.[T][?]this
-_______________________________
-_______________________________
-Bye. Hope to see you again soon!
-_______________________________
-
